bpf: try to reduce the number of generated insns

+++ bpf/common/strings.h
@@ -8,7 +8,7 @@ static __always_inline char lowercase(char c) {
 }
 
 static __always_inline bool stricmp(const char *s1, const char *s2, u8 n) {
-    for (u16 i = 0; i < n && s1[i] && s2[i]; i++) {
+    for (u8 i = 0; i < n && s1[i] && s2[i]; i++) {
         if (lowercase(s1[i]) != lowercase(s2[i])) {
             return false;
         }
